+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Fjor Changelog
 ==============
 
+*	0.1.2
+
+	*	More easy to instantiate using utility static method.
+	*	Pluggable (`AfterNew` event).
+
 *	0.1.1
 
 	*	Now installable through Composer.